pyproject.toml
[project]
name = "hft"
version = "0.1.0"
description = ""
authors = [
    {name = "paku",email = "htphee.paku@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "websockets (>=15.0.1,<16.0.0)",
    "nats-py (>=2.10.0,<3.0.0)",
    "colorlog (>=6.9.0,<7.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "motor (>=3.7.1,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "notebook (>=7.4.3,<8.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "numba (>=0.61.2,<0.62.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "beautifulsoup4 (>=4.12.0,<5.0.0)",
    "holidays (>=0.81,<0.82)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-asyncio = "^1.2.0"
pytest-mock = "^3.15.0"
pytest-cov = "^7.0.0"

